% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_plot.R
\name{finalise_plot}
\alias{finalise_plot}
\title{Arrange alignment and save ggplot chart}
\usage{
finalise_plot(
  plot_name,
  source_name,
  save_filepath = file.path(Sys.getenv("TMPDIR"), "tmp-nc.png"),
  width_pixels = 720,
  height_pixels = 405
)
}
\arguments{
\item{plot_name}{The variable name of the plot you have created that you want to format and save}

\item{source_name}{Text for left hand side of the plot footnote}

\item{save_filepath}{Exact filepath that you want the plot to be saved to}

\item{width_pixels}{Width in pixels that you want to save your chart to - defaults to 720}

\item{height_pixels}{Height in pixels that you want to save your chart to - defaults to 405}
}
\value{
(Invisibly) an updated ggplot object.
}
\description{
Arrange alignment and save ggplot chart
}
\examples{
finalise_plot(plot_name = myplot,
source = "The source for my data",
save_filepath = "filename_that_my_plot_should_be_saved_to-nc.png",
width_pixels = 720,
height_pixels = 405,
logo_image_path = "logo_image_filepath.png"
)

}
\keyword{finalise_plot}
