% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_cols.R
\name{ow_cols}
\alias{ow_cols}
\title{Function to extract useful colors as hex codes}
\usage{
ow_cols(...)
}
\arguments{
\item{...}{Character names of ow_colors}
}
\description{
Function to extract useful colors as hex codes
}
\examples{

ow_cols()
ow_cols("labour")
ow_cols("conservative", "lab", "greens")

}
