% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_cols.R
\name{ow_pal}
\alias{ow_pal}
\title{Return function to interpolate a ow color palette}
\usage{
ow_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ow_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a ow color palette
}
